+++ .env
@@ -1,0 +1,6 @@
+REVIEWED_TRANSLATIONS=ja,en
+NOT_STARTED_TRANSLATIONS=hu,ru
+IN_REVIEW_TRANSLATIONS=none
+DEFAULT_LOCALE=en
+LOCALES=ja,hu,ru,en
+ENV=development
