+++ .gitignore
@@ -27,7 +27,7 @@ yarn-error.log*
 
 # local env files
 .env*.local
-.env
+!.env
 
 # vercel
 .vercel
